pyproject.toml
[project]
name = "xarray-tropical-cyclones"
version = "0.1.0"
description = "Revisiting tropical cyclones IBTrACS data with xarray."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair-tiles>=0.3.0",
    "altair>=5.4.1",
    "cartopy>=0.24.1",
    "h5netcdf>=1.4.0",
    "matplotlib>=3.9.2",
    "python-dotenv>=1.0.1",
    "vl-convert-python>=1.7.0",
    "xarray>=2024.10.0",
]
